use `count` function instead of listing all messages
`Message.objects.all()` will cause Django to retrieve list of all messages, create list of messages objects and return them. This is not optimal when we just need to to check if object exists or not. best option is using `exists` function but its not supported by django versions < 1.6 so to be safe I am suggesting `count`